light_client::testing

Function qc_chain_from_leaf_chain

pub fn qc_chain_from_leaf_chain<'a>(
    leaves: impl IntoIterator<Item = &'a LeafQueryData<SeqTypes>>,
) -> Vec<Certificate> 

Extract a chain of QCs from a chain of leaves.

The resulting QC chain will be one shorter than the leaf chain, and will justify the finality of the leaf preceding this leaf chain, since we extract QCs from the justifying QC of each leaf.