light_client::consensus::quorum

Trait StakeTablePair

pub trait StakeTablePair {
    // Required methods
    fn stake_table(
        &self,
    ) -> impl Send + Future<Output = Result<Arc<StakeTable>>>;
    fn next_epoch_stake_table(
        &self,
    ) -> impl Send + Future<Output = Result<Arc<StakeTable>>>;
}

Getters for the current epoch’s stake table and the next.

The current stake_table is always needed to verify a Certificate from this epoch. Depending on the Certificate, the next epoch’s stake table may also need to be fetched (in the case where the certificate is part of an epoch transition).

Required Methods

fn stake_table(&self) -> impl Send + Future<Output = Result<Arc<StakeTable>>>

Get the stake table for the current epoch.

fn next_epoch_stake_table( &self, ) -> impl Send + Future<Output = Result<Arc<StakeTable>>>

Get the stake table for the next epoch.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.