light_client

Module testing

Structs

AlwaysFalseQuorum

AlwaysTrueQuorum

EpochChangeQuorum

A quorum which verifies that epoch change QCs are provided, but does not check signatures.

InnerTestClient 🔒

TestClient

VersionCheckQuorum

A quorum which verifies that calls to verify use the correct version, but does not check signatures.

Functions

custom_epoch_change_leaf_chain

Construct a customized leaf chain during which the epoch advances.

custom_leaf_chain

Construct a customized leaf chain for the given height range.

custom_leaf_chain_with_upgrade

Construct a customized leaf chain for the given height range.

epoch_change_leaf_chain

Construct a valid leaf chain during which the epoch advances.

leaf_chain

Construct a valid leaf chain for the given height range.

leaf_chain_with_upgrade

Construct a valid leaf chain for the given height range.

qc_chain_from_leaf_chain

Extract a chain of QCs from a chain of leaves.

random_validator

register_validator_events 🔒

Type Aliases

EnableEpochs

Upgrade to epochs during testing.

LegacyVersion

Test without epochs and with legacy HotStuff.